chore: cleanup unused constants
We have some leftover constants from some previous changes. This
just cleans up all that are no longer needed.

+++ s3api/utils/chunk-reader.go
@@ -174,7 +174,7 @@ func NewChunkReader(ctx *fiber.Ctx, r io.Reader, authdata AuthData, region, secr
 	}
 
 	if decContLength > maxObjSizeLimit {
-		debuglogger.Logf("the object size exceeds the allowed limit: (size): %v, (limit): %v", decContLength, maxObjSizeLimit)
+		debuglogger.Logf("the object size exceeds the allowed limit: (size): %v, (limit): %v", decContLength, int64(maxObjSizeLimit))
 		return nil, s3err.GetAPIError(s3err.ErrEntityTooLarge)
 	}
